Completed upgrade of selection=11 case to allow variable and distributed encoding, compactly enough to fit on 64 bit gene. Coding is not position dependent : each byte may code for a set of LUT rules:   s allowed from 1-7 bit 7                              bit 6        bit 5        bit 4        bit 3                        bit 2                       bit 1                          bit    0 birth/survival                 s bit 2    s bit 1     s bit 0      allow se-s0==3      allow se-s0==2      allow se-s0==1        allow se-s0==0
the missing case s==4 and se==4 is coded in the free codes  x0000001  with x = 1 for birth and 0 for survival i.e. 0x81 and 0x01.
i.e. to turn on birth for all 5 s==4 LUT cases se 0 to 4 we need to specify two bytes :  cf  and 81 somewhere in the gene.